+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.7.0] – 2019-12-03
+
+- added support for Symfony 5.*
+- dropped support for Symfony 4.1
+
+[2.7.0]: https://github.com/craue/TwigExtensionsBundle/compare/2.6.0...2.7.0
+
 ## [2.6.0] – 2019-11-14
 
 - deprecated passing the service container to `setLocale`, pass the request stack instead